feat(history): get bid history of old auctions
