+++ src/app/components/color-history-modal/color-history-modal.component.ts
@@ -1,104 +1,8 @@
-import { HttpClient } from '@angular/common/http'
-import { Component, OnInit, ViewContainerRef } from '@angular/core'
+import { Component, OnInit } from '@angular/core'
 import { BsModalRef, BsModalService } from 'ngx-bootstrap/modal'
-import { NgxChartsModule } from '@swimlane/ngx-charts'
 
 import { Color } from 'src/app/services/store/store.service'
-import { environment } from 'src/environments/environment'
-import { ApiService } from 'src/app/services/api/api.service'
-
-export interface Result {
-  consumed_gas: number
-  storage_size: number
-  paid_storage_size_diff?: number
-}
-
-export interface Child4 {
-  prim: string
-  type: string
-  name: string
-  value: string
-}
-
-export interface Child3 {
-  prim: string
-  type: string
-  children: Child4[]
-}
-
-export interface Child2 {
-  prim: string
-  type: string
-  value: string
-  name: string
-  children: Child3[]
-}
-
-export interface Child {
-  prim: string
-  type: string
-  name: string
-  value: string
-  children: Child2[]
-}
-
-export interface Parameters {
-  prim: string
-  type: string
-  value: string
-  children: Child[]
-  name: string
-}
-
-export interface Child6 {
-  prim: string
-  type: string
-  name: string
-  value: any
-  from: string
-  diff_type: string
-}
-
-export interface Child5 {
-  prim: string
-  type: string
-  name: string
-  children: Child6[]
-  value: any
-  diff_type: string
-}
-
-export interface StorageDiff {
-  prim: string
-  type: string
-  children: Child5[]
-}
-
-export interface HistoryItem {
-  level: number
-  fee: number
-  counter: number
-  gas_limit: number
-  amount: number
-  content_index: number
-  result: Result
-  parameters: Parameters
-  storage_diff: StorageDiff
-  timestamp: Date
-  id: string
-  protocol: string
-  hash: string
-  network: string
-  kind: string
-  source: string
-  destination: string
-  status: string
-  entrypoint: string
-  internal: boolean
-  mempool: boolean
-  storage_limit?: number
-  burned?: number
-}
+import { ApiService, HistoryItem } from 'src/app/services/api/api.service'
 
 @Component({
   selector: 'app-auction-modal',
@@ -120,7 +24,6 @@ export class ColorHistoryModalComponent implements OnInit {
   constructor(
     public bsModalRef: BsModalRef,
     public modalService: BsModalService,
-    private readonly http: HttpClient,
     private readonly api: ApiService
   ) {}
 
@@ -135,11 +38,9 @@ export class ColorHistoryModalComponent implements OnInit {
 
   public async getHistory() {
     if (this.color && this.color.auction) {
-      this.history = await this.http
-        .get<HistoryItem[]>(
-          `${environment.indexerUrl}auction/operations?entrypoint=bid&parameters.value=${this.color.auction.auctionId}`
-        )
-        .toPromise()
+      this.history = await this.api.getBidsForAuction(
+        this.color.auction.auctionId
+      )
     }
   }
 
@@ -158,6 +59,12 @@ export class ColorHistoryModalComponent implements OnInit {
         a.maxBid = maxBids[a.parameters.children[0].value] ?? 0
         a.ask = a.parameters.children[1].value
         return a
+      })
+
+      auctions.forEach(async (a: any) => {
+        a.bids = await this.api.getBidsForAuction(
+          a.parameters.children[0].value
+        )
       })
 
       const series = auctions
